pyproject.toml
[project]
name = "jira-commit-msg"
version = "0.1.0"
description = "A pre-commit compliant hook to add JIRA ticket data to commit msg"
authors = [{ name = "Louis-Amaury Chaib", email = "louisamaury.chaib@gmail.com" }]
requires-python = ">=3.10,<4"
readme = "README.md"
license = "MIT"
classifiers = [
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Programming Language :: Python :: 3.14",
]
dependencies = [
    "typer>=0.21.1,<0.22",
    "gitpython>=3.1.46,<4",
]

[project.scripts]
jira-commit-msg = "jira_commit_msg:app"

[dependency-groups]
dev = [
    "pytest>=8.4.2,<9",
    "mypy>=1.19.1,<2",
    "ruff>=0.14.13,<0.15",
    "pytest-coverage>=0.0,<0.1",
    "pre-commit>=4.3.0,<5",
]


[build-system]
requires = ["uv_build>=0.9.26,<0.10.0"]
build-backend = "uv_build"

[tool.pytest.ini_options]
addopts = "-vv"
testpaths = [
    "."
]

[tool.coverage.paths]
sources = ["src/"]

[tool.coverage.run]
relative_files = true

[tool.coverage.report]
show_missing = true
fail_under = 100
skip_empty = true
skip_covered = true

[tool.ruff]
line-length = 100

[tool.ruff.lint]
select = ["ALL"]
ignore = ["D", "FBT", "COM812"]

[tool.ruff.lint.per-file-ignores]
"test/**" = [
    "S101", # use of assert
    "ANN201",  # missing return type annotations,
    
    "INP001", 
]
